# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Olimpiadas",
        "description": "Documentación de la API del sistema de Olimpiadas",
        "contact": {
            "email": "admin@olimpiadas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/boleta/{registro}": {
            "get": {
                "tags": [
                    "Boletas"
                ],
                "summary": "Generar PDF de boleta",
                "operationId": "f856c3b1a8d05a81acd84e8ad72d2307",
                "parameters": [
                    {
                        "name": "registro",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF de boleta generado exitosamente",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                }
            }
        },
        "/api/boleta/enviar": {
            "post": {
                "tags": [
                    "Boletas"
                ],
                "summary": "Enviar boleta por correo",
                "operationId": "7abd68af44024a46ef05a8e442896956",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "boleta_id",
                                    "email"
                                ],
                                "properties": {
                                    "boleta_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@ejemplo.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Boleta enviada exitosamente"
                    }
                }
            }
        },
        "/api/boleta/ocr": {
            "post": {
                "tags": [
                    "Boletas"
                ],
                "summary": "Extraer número desde OCR",
                "operationId": "65700a982aa3bdbc1d2ef825c91d734e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "imagen": {
                                        "description": "Imagen del comprobante",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Número extraído exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "numero": {
                                            "type": "string",
                                            "example": "123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/boleta/ocr-imagen": {
            "post": {
                "tags": [
                    "Boletas"
                ],
                "summary": "Procesar imagen con OCR",
                "operationId": "0e76d1417a9400ed6e1e0f6384cd01f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "imagen": {
                                        "description": "Imagen a procesar",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Imagen procesada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "texto": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/inscripcion/tutor": {
            "post": {
                "tags": [
                    "Tutores"
                ],
                "summary": "Registrar nuevo tutor",
                "description": "Registra un nuevo tutor en el sistema",
                "operationId": "a204a0ebdfd3c776a4fc29de0d673c46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombres",
                                    "apellidos",
                                    "documento_identidad",
                                    "telefono",
                                    "correo_electronico",
                                    "direccion"
                                ],
                                "properties": {
                                    "nombres": {
                                        "type": "string",
                                        "example": "María"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "González"
                                    },
                                    "documento_identidad": {
                                        "type": "string",
                                        "example": "87654321"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "77712345"
                                    },
                                    "correo_electronico": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@ejemplo.com"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Principal #123"
                                    },
                                    "competidores": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "es_principal": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "relacion": {
                                                    "type": "string",
                                                    "example": "madre"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tutor registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tutor registrado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Tutor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Tutor": {
                "title": "Tutor",
                "description": "Modelo de Tutor",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "nombres": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "documento_identidad": {
                        "type": "string",
                        "example": "123456"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "77712345"
                    },
                    "correo_electronico": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@ejemplo.com"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Calle 123"
                    },
                    "competidores": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "es_principal": {
                                    "type": "boolean"
                                },
                                "relacion": {
                                    "type": "string",
                                    "example": "padre"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Ingresa tu token JWT (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}